feat: refactor ast-node
